backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
qdrant-client==1.7.3
pymupdf==1.23.8
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-docx==1.1.0
